Restructure Components organization
